Update Recognize.py
Initially, the program would often mistakenly identify me as someone else, and therefore, would save their name instead of mine. The program now displays the confidence factor below the person's face and only if conf > 68 will it make the attendance, else it would not register it. Removed the Unknown images, as it was heavy on the PC and had no actual requirement. Added few more flags and factors, so that face recognition is stricter and made the name displayed on top remove the [] brackets and commas, so it displays only the name and ID.